Refactor services to share common code
