+++ app/res/layout/list_item_app.xml
@@ -86,7 +86,7 @@
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
         android:layout_below="@+id/itemAppGrammar"
-        android:text="@string/defaultServerHttp"
+        android:text="@string/defaultHttpServer"
         android:textColor="@color/orange"
         android:textSize="14sp" />
