+++ app/res/values/keys.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="prefCurrentSortOrder" translatable="false">CurrentSortOrder</string>
-    <string name="keyServerHttp" translatable="false">keyServerHttp</string>
-    <string name="keyServerWs" translatable="false">keyServerWs</string>
+    <string name="keyHttpServer" translatable="false">keyHttpServer</string>
+    <string name="keyWsServer" translatable="false">keyWsServer</string>
     <string name="keyRecordingRate" translatable="false">keyRecordingRate</string>
     <string name="keyAutoStart" translatable="false">keyAutoStart</string>
     <string name="keyAutoStopAfterPause" translatable="false">keyAutoStopAfterPause</string>
@@ -18,10 +18,26 @@
     <string name="keyImeHelpText" translatable="false">keyImeHelpText</string>
     <string name="keyImeCombo" translatable="false">keyImeCombo</string>
     <string name="keyImeCurrentCombo" translatable="false">keyImeCurrentCombo</string>
-    <string name="defaultServerHttp" translatable="false">http://bark.phon.ioc.ee/speech-api/v1/recognize</string>
+
+    // ws://bark.phon.ioc.ee:82/dev/duplex-speech-api/ws/speech
+    // ws://bark.phon.ioc.ee:82/dev/duplex-speech-api/ws/status
+    <string name="defaultWsServer" translatable="false">ws://bark.phon.ioc.ee:82/dev/duplex-speech-api/ws/speech</string>
+
+    <string name="defaultHttpServer" translatable="false">http://bark.phon.ioc.ee/speech-api/v1/recognize</string>
     // These keys are needed to locate (to show/hide) the "enable ime" link in the settings
     <string name="keyEnableIme" translatable="false">keyEnableIme</string>
     <string name="keyCategoryIme" translatable="false">keyCategoryIme</string>
+
+    <string name="defaultAutoStopAfterTime" translatable="false">20</string>
+
+    <!-- TODO: convert to integer-array -->
+    <string-array name="valuesAutoStopAfterTime" translatable="false">
+        <item>2</item>
+        <item>5</item>
+        <item>10</item>
+        <item>20</item>
+        <item>30</item>
+    </string-array>
 
     <array name="keysActivity" translatable="false">
         <item>@string/keyCombo</item>
