+++ app/res/xml/preferences_server_ws.xml
@@ -3,10 +3,10 @@
     // Settings that apply to the WebSocket-based server (also denoted as "fast" and "continuous")
     // These cannot be configured using standard extras
     <EditTextPreference
-        android:defaultValue="@string/defaultServerWs"
+        android:defaultValue="@string/defaultWsServer"
         android:inputType="textUri"
-        android:key="@string/keyServerWs"
-        android:summary="@string/defaultServerWs"
+        android:key="@string/keyWsServer"
+        android:summary="@string/defaultWsServer"
         android:title="@string/titleDefaultServiceWs"></EditTextPreference>
     <CheckBoxPreference
         android:defaultValue="@bool/defaultImeAutoStopAfterPause"
